cleanup

+++ modules/mvstore/src/main/scala/org/ergoplatform/uexplorer/mvstore/SuperNodeCollector.scala
@@ -11,11 +11,11 @@ import java.util.zip.GZIPInputStream
 import scala.collection.concurrent
 import scala.io.Source
 import scala.jdk.CollectionConverters.*
-import scala.util.{Random, Try}
+import scala.util.{Random, Success, Try}
 
 class SuperNodeCollector[HK: HotKeyCodec](superNodeFile: File) extends LazyLogging {
   private val hotKeyCodec: HotKeyCodec[HK] = implicitly[HotKeyCodec[HK]]
-  private val hotKeyFileNameGz            = "hot-keys.csv.gz"
+  private val hotKeyFileNameGz             = "hot-keys.csv.gz"
 
   private lazy val stringifiedHotKeys: Map[HK, String] =
     Source
@@ -46,38 +46,41 @@ class SuperNodeCollector[HK: HotKeyCodec](superNodeFile: File) extends LazyLoggi
 
   def getHotKeyString(hotKey: HK): Option[String] = stringifiedHotKeys.get(hotKey)
 
-  def report(hotKeysWithCounter: Iterator[(HK, Counter)]): Try[String] = {
-    val keysByCount =
-      hotKeysWithCounter
-        .collect {
-          case (hotKey, counter) if counter.isHot && !stringifiedHotKeys.contains(hotKey) =>
-            hotKeyCodec.serialize(hotKey) -> counter
-        }
-        .toVector
-        .sortBy(_._2.writeOps)(Ordering[Long].reverse)
-        .map { case (hotKeyString, Counter(writeOps, readOps, boxesAdded, boxesRemoved)) =>
-          val stats  = s"$writeOps $readOps $boxesAdded $boxesRemoved"
-          val indent = 48
-          s"$stats ${List.fill(Math.max(1, indent - stats.length))(" ").mkString("")} $hotKeyString"
-        }
-    logger.info(s"Collected ${keysByCount.size} supernodes, writing to ${superNodeFile.getAbsolutePath} : ")
-    Try {
-      keysByCount.headOption
-        .map { _ =>
-          logger.info(s"Writing hot keys to ${superNodeFile.getAbsolutePath}")
-          val report     = keysByCount.mkString("", "\n", "\n")
-          val fileWriter = new FileWriter(superNodeFile)
-          try fileWriter.write(report)
-          finally fileWriter.close()
-          report
-        }
-        .getOrElse("")
+  def writeReport(hotKeysWithCounter: Iterator[(HK, Counter)]): Try[_] =
+    if (superNodeFile.exists()) {
+      logger.info(s"Skipping report as file ${superNodeFile.getAbsolutePath} already exists")
+      Success(())
+    } else {
+      val keysByCount =
+        hotKeysWithCounter
+          .collect {
+            case (hotKey, counter) if counter.isHot && !stringifiedHotKeys.contains(hotKey) =>
+              hotKeyCodec.serialize(hotKey) -> counter
+          }
+          .toVector
+          .sortBy(_._2.writeOps)(Ordering[Long].reverse)
+          .map { case (hotKeyString, Counter(writeOps, readOps, boxesAdded, boxesRemoved)) =>
+            val stats  = s"$writeOps $readOps $boxesAdded $boxesRemoved ${boxesAdded - boxesRemoved}"
+            val indent = 30
+            s"$stats ${List.fill(Math.max(4, indent - stats.length))(" ").mkString("")} $hotKeyString"
+          }
+      Try {
+        keysByCount.headOption
+          .map { _ =>
+            logger.info(s"Writing ${keysByCount.size} hot keys to ${superNodeFile.getAbsolutePath}")
+            val report     = keysByCount.mkString("", "\n", "\n")
+            val fileWriter = new FileWriter(superNodeFile)
+            try fileWriter.write(report)
+            finally fileWriter.close()
+            report
+          }
+          .getOrElse("")
+      }
     }
-  }
 }
 
 object SuperNodeCollector {
-  private val hotLimit = 5000
+  private val hotLimit = 10000
   case class Counter(writeOps: Long, readOps: Long, boxesAdded: Int, boxesRemoved: Int) {
     def this() = this(0, 0, 0, 0)
     def isHot: Boolean = writeOps > hotLimit || readOps > hotLimit || boxesAdded > hotLimit || boxesRemoved > hotLimit
